LayoutTests/inspector/dom-debugger/url-breakpoints-dom-containing.html

<!doctype html>
<html>
<head>
<script src="../../http/tests/inspector/resources/inspector-test.js"></script>
<script src="../debugger/resources/breakpoint-options-utilities.js"></script>
<script src="../debugger/resources/log-active-stack-trace.js"></script>
<script src="../worker/resources/resource-utilities.js"></script>
<script src="resources/url-breakpoint-utilities.js"></script>
<script>
function test()
{
    const type = WI.URLBreakpoint.Type.Text;
    const url = "data";

    let suite = InspectorTest.createAsyncSuite("URLBreakpoint.BreakOnContainingText");

    suite.addTestCase({
        name: "URLBreakpoint.BreakOnContainingText.DOM",
        description: "Check that debugger pauses on a URL containing text.",
        async test() {
            let breakpoint = await InspectorTest.URLBreakpoint.createBreakpoint(type, url);

            InspectorTest.URLBreakpoint.request("DOM");
            await WI.debuggerManager.awaitEvent(WI.DebuggerManager.Event.Paused);

            let targetData = WI.debuggerManager.dataForTarget(WI.debuggerManager.activeCallFrame.target);
            InspectorTest.log("PAUSED:");
            InspectorTest.expectEqual(targetData.pauseReason, WI.DebuggerManager.PauseReason.URL, `Pause reason should be URL.`);
            InspectorTest.expectEqual(targetData.pauseData.breakpointURL, breakpoint.url, "Pause data URL should match breakpoint URL.");

            logActiveStackTrace();

            await WI.debuggerManager.resume();
        },
        teardown: InspectorTest.URLBreakpoint.teardown,
    });

    InspectorTest.BreakpointOptions.addTestCases(suite, {
        testCaseNamePrefix: "DOM.",
        createBreakpoint() {
            return InspectorTest.URLBreakpoint.createBreakpoint(type, url);
        },
        triggerBreakpoint() {
            return InspectorTest.URLBreakpoint.request("DOM");
        },
    });

    suite.runTestCasesAndFinish();
}
</script>
</head>
<body onload="runTest()">
<p>Tests for text URL breakpoints.</p>
</body>
</html>
