LayoutTests/inspector/dom-debugger/url-breakpoints-dom-containing-expected.txt

CONSOLE MESSAGE: BREAKPOINT ACTION LOG 1
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 2
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 3
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 4
Tests for text URL breakpoints.


== Running test suite: URLBreakpoint.BreakOnContainingText
-- Running test case: URLBreakpoint.BreakOnContainingText.DOM
Adding "text:data" URL Breakpoint...
Request DOM "resources/dataDOM.json"...
PAUSED:
PASS: Pause reason should be URL.
PASS: Pause data URL should match breakpoint URL.
CALL STACK:
0: [F] loadResourceDOM
1: [P] Global Code
-- Running test teardown.

-- Running test case: URLBreakpoint.BreakOnContainingText.DOM.Options.Condition
Adding "text:data" URL Breakpoint...

Setting condition to 'false'...

Triggering breakpoint...
Request DOM "resources/dataDOM.json"...
PASS: Should not pause.

Triggering breakpoint...
Request DOM "resources/dataDOM.json"...
PASS: Should not pause.

Setting condition to 'true'...

Triggering breakpoint...
Request DOM "resources/dataDOM.json"...
PASS: Should pause.

Triggering breakpoint...
Request DOM "resources/dataDOM.json"...
PASS: Should pause.

-- Running test case: URLBreakpoint.BreakOnContainingText.DOM.Options.Condition.ConsoleCommandLineAPI
Adding saved console value 'false'...
Adding "text:data" URL Breakpoint...

Setting condition to saved console value...

Triggering breakpoint...
Request DOM "resources/dataDOM.json"...
PASS: Should not pause.

Triggering breakpoint...
Request DOM "resources/dataDOM.json"...
PASS: Should not pause.

Adding saved console value 'true'...
Setting condition to saved console value...

Triggering breakpoint...
Request DOM "resources/dataDOM.json"...
PASS: Should pause.

Triggering breakpoint...
Request DOM "resources/dataDOM.json"...
PASS: Should pause.

-- Running test case: URLBreakpoint.BreakOnContainingText.DOM.Options.IgnoreCount
Adding "text:data" URL Breakpoint...

Setting ignoreCount to '2'...

Triggering breakpoint...
Request DOM "resources/dataDOM.json"...
PASS: Should not pause.

Triggering breakpoint...
Request DOM "resources/dataDOM.json"...
PASS: Should not pause.

Triggering breakpoint...
Request DOM "resources/dataDOM.json"...
PASS: Should pause.

Triggering breakpoint...
Request DOM "resources/dataDOM.json"...
PASS: Should pause.

-- Running test case: URLBreakpoint.BreakOnContainingText.DOM.Options.Action.Log
Adding "text:data" URL Breakpoint...

Adding log action...

Triggering breakpoint...
Request DOM "resources/dataDOM.json"...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...

Triggering breakpoint...
Request DOM "resources/dataDOM.json"...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...
Enabling auto-continue...

Triggering breakpoint...
Request DOM "resources/dataDOM.json"...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing log action...

Triggering breakpoint...
Request DOM "resources/dataDOM.json"...
PASS: Should execute breakpoint action.
PASS: Should not pause.

-- Running test case: URLBreakpoint.BreakOnContainingText.DOM.Options.Actions.Evaluate
Adding "text:data" URL Breakpoint...

Adding evaluate action...

Triggering breakpoint...
Request DOM "resources/dataDOM.json"...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...

Triggering breakpoint...
Request DOM "resources/dataDOM.json"...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...
Enabling auto-continue...

Triggering breakpoint...
Request DOM "resources/dataDOM.json"...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing evaluate action...

Triggering breakpoint...
Request DOM "resources/dataDOM.json"...
PASS: Should execute breakpoint action.
PASS: Should not pause.

-- Running test case: URLBreakpoint.BreakOnContainingText.DOM.Options.Actions.Evaluate.ConsoleCommandLineAPI
Adding saved console value '1'...
Adding "text:data" URL Breakpoint...

Adding evaluate action using saved console value...

Triggering breakpoint...
Request DOM "resources/dataDOM.json"...
PASS: Should execute breakpoint action.
PASS: Should pause.

Adding saved console value '2'...
Editing evaluate action using saved console value...

Triggering breakpoint...
Request DOM "resources/dataDOM.json"...
PASS: Should execute breakpoint action.
PASS: Should pause.

Adding saved console value '3'...
Editing evaluate action using saved console value...
Enabling auto-continue...

Triggering breakpoint...
Request DOM "resources/dataDOM.json"...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Adding saved console value '4'...
Editing evaluate action using saved console value...

Triggering breakpoint...
Request DOM "resources/dataDOM.json"...
PASS: Should execute breakpoint action.
PASS: Should not pause.
